requirements.txt.txt
transformers
scikit-image
Pillow
matplotlib
numpy
scipy
requests
opencv-python
torch==1.13.1+cu116 torchvision==0.14.1+cu116 torchaudio==0.13.1 --extra-index-url https://download.pytorch.org/whl/cu116